requirements.txt
# =====================================
# ✅ Core Framework
# =====================================
fastapi==0.116.1
uvicorn==0.35.0
starlette==0.47.1

# =====================================
# ✅ Gmail API + OAuth 2.0 Integration
# =====================================
google-api-python-client==2.176.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
httplib2==0.22.0
oauthlib==3.3.1
requests-oauthlib==2.0.0
uritemplate==4.2.0
rsa==4.9.1
pyasn1==0.6.1
pyasn1_modules==0.4.2

# =====================================
# ✅ Background Scheduling
# =====================================
APScheduler==3.11.0
tzlocal==5.3.1

# =====================================
# ✅ Database Connector (MySQL)
# =====================================
mysql-connector-python==9.3.0

# =====================================
# ✅ Transformers + Model Inference
# =====================================
openai==1.95.1     # Remove if using FLAN-T5 locally without OpenAI API

# =====================================
# ✅ Helper Libraries
# =====================================
requests==2.32.4
httpx==0.28.1
httpcore==1.0.9
h11==0.16.0
typing-extensions==4.14.1
typing-inspection==0.4.1
annotated-types==0.7.0
anyio==4.9.0
distro==1.9.0
charset-normalizer==3.4.2
idna==3.10
certifi==2025.7.9
python-dotenv==1.1.1
protobuf==6.31.1
proto-plus==1.26.1
googleapis-common-protos==1.70.0
cachetools==5.5.2

# =====================================
# ✅ Development & Debugging Tools
# =====================================
tqdm==4.67.1
jiter==0.10.0
click==8.2.1
pyparsing==3.2.3
pydantic==2.11.7
pydantic_core==2.33.2
